refactor: 충돌 해결 및 에러 레벨 추가
